+++ forge/src/main/java/jimenezli/neuro21/forge/Neuro21ModForge.java
@@ -8,6 +8,7 @@ import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 @Mod(Neuro21Mod.MOD_ID)
@@ -23,10 +24,15 @@ public class Neuro21ModForge {
         }
 
         bus.addListener(this::clientSetup);
+        bus.addListener(this::commonSetup);
         MinecraftForge.EVENT_BUS.register(this);
     }
 
     private void clientSetup(final FMLClientSetupEvent event) {
         Neuro21ModClient.Init();
+    }
+
+    private void commonSetup(final FMLCommonSetupEvent event) {
+        event.enqueueWork(Neuro21Mod::Init);
     }
 }